+++ docs/versionhistory.rst
@@ -49,6 +49,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
   arrives in an exception group)
 - Fixed support for Linux abstract namespaces in UNIX sockets that was broken in v4.2
   (#781 <https://github.com/agronholm/anyio/issues/781>_; PR by @tapetersen)
+- Fixed asyncio task groups not yielding control to the event loop at exit if there were
+  no child tasks to wait on
 
 **4.4.0**